Removed old types
