+++ src/index.ts
@@ -1,24 +1,1 @@
-export {
-	AnchorType,
-	CenterType,
-	Fill,
-	Left,
-	Right,
-	Top,
-	Bottom,
-	LeftResizable,
-	RightResizable,
-	TopResizable,
-	BottomResizable,
-	Positioned,
-	Custom,
-	Fixed,
-	ViewPort,
-	Layer,
-	Centered,
-	CenteredVertically,
-} from "./components";
-
-//export { useParentSpace } from "./hooks/useParentSpace";
-
-//export { ResizeType, ISpaceInfo, IParentSpace, IPosition } from "./types";
+export * from "./components";
